Convert Google Maps logic to a behavior.
