+++ src/behavior/index.js
@@ -1,10 +1,12 @@
 // Import pre-defined behaviors from src/behavior.
 import edit from './edit';
+import google from './google';
 import measure from './measure';
 import rememberLayer from './rememberLayer';
 
 export default {
   edit,
+  google,
   measure,
   rememberLayer,
 };
